# backend/api/swagger.yaml
definitions:
  api.ChallengeFriend.challengeFriendRequest:
    properties:
      friend_id:
        type: integer
      settings:
        $ref: '#/definitions/model.DraftSettings'
    type: object
...
  api.GetCards.getCardResponse:
    properties:
      card_ids:
        items:
          type: integer
        type: array
      number:
        type: integer
    type: object
  api.GetChallenges.getChallengesResponse:
    properties:
      challenges:
        items:
          $ref: '#/definitions/model.DraftChallenge'
        type: array
    type: object
...
  api.GetRandomCards.getRandomCardsResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/model.Card'
        type: array
    type: object
...
  model.Card:
    properties:
      atk:
        type: integer
      attribute:
        type: string
      card_images:
        items:
          $ref: '#/definitions/model.CardImage'
        type: array
      card_sets:
        items:
          $ref: '#/definitions/model.CardSet'
        type: array
      def:
        type: integer
      desc:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      race:
        type: string
      sets:
        type: string
      type:
        type: string
    type: object
  model.CardImage:
    properties:
      id:
        type: integer
      image_url:
        type: string
      image_url_small:
        type: string
    type: object
  model.CardSet:
    properties:
      set_code:
        type: string
      set_name:
        type: string
      set_rarity:
        type: string
      set_rarity_code:
        type: string
    type: object
  model.DraftChallenge:
    properties:
      challenge_date:
        type: string
      challenger_id:
        type: integer
      id:
        type: integer
      receiver_id:
        type: integer
      settings:
        $ref: '#/definitions/model.DraftSettings'
      status:
        type: string
    type: object
  model.DraftMode:
    enum:
    - bestof
    - round
    type: string
    x-enum-varnames:
    - DraftModeBestOf
    - DraftGoalRounds
  model.DraftSettings:
    properties:
      extra_deck_draws:
        type: integer
      main_deck_draws:
        type: integer
      mode:
        $ref: '#/definitions/model.DraftMode'
      modeValue:
        type: integer
      sets:
        items:
          $ref: '#/definitions/model.CardSet'
        type: array
    type: object
...
info:
  contact: {}
paths:
  /cards:
    get:
      description: Endpoint used to retrieve all cards.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetCards.getCardResponse'
        "500":
          description: Internal Server Error
      summary: Endpoint used to retrieve all cards.
      tags:
      - Yu-Gi-Oh
  /cards/{cardID}:
    get:
      description: Endpoint used to retrieve a specific cards.
      parameters:
      - description: The id of the requested card.
        in: path
        name: cardID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Card'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Endpoint used to retrieve a specific card.
      tags:
      - Yu-Gi-Oh
  /cards/random:
    get:
      description: |-
        Retrieve a certain amount of random card at once. Calling the endpoint without any parameters returns
        a only one random card.
      parameters:
      - description: Determines the amount of random cards to be returned.
        in: query
        name: size
        type: integer
      - description: Contains the filtered sets, separated by comma, e.g., 'Set 1,Set
          2,Set 3'
        in: query
        name: sets
        type: string
      - description: Contains the filtered types, separated by comma, e.g., 'Fire,
          Water, Dragon'
        in: query
        name: types
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetRandomCards.getRandomCardsResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Retrieve a certain amount of random card at once.
      tags:
      - Yu-Gi-Oh
  /drafts/challenges:
    get:
      description: Retrieve the challenges of the current user.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetChallenges.getChallengesResponse'
        "401":
          description: Unauthorized.
          schema:
            type: string
        "500":
          description: Internal Server Error.
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Retrieve the challenges of the current user.
      tags:
      - Draft
    post:
      consumes:
      - application/json
      description: Challenge a friend to a draft.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      - description: Contains the information for the receiving party.
        in: body
        name: receiver
        required: true
        schema:
          $ref: '#/definitions/api.ChallengeFriend.challengeFriendRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Body data is not correct/valid
          schema:
            type: string
        "500":
          description: Internal Server Error.
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Challenge a friend to a draft.
      tags:
      - Draft
  /drafts/challenges/{id}/accept:
    post:
      description: Accept a draft challenge from another user.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      - description: Contains the id of the challenge to be accepted.
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Body data is not correct/valid
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "403":
          description: Thrown when anyone except the receiver tries to accept a challenge.
          schema:
            type: string
        "500":
          description: Internal Server Error.
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Accept a draft challenge from another user.
      tags:
      - Draft
  /drafts/challenges/{id}/decline:
    post:
      description: Decline a draft challenge from another user.
      parameters:
      - description: Contains the authorization token.
        in: header
        name: authorization
        required: true
        type: string
      - description: Contains the id of the challenge to be declined.
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Body data is not correct/valid
          schema:
            type: string
        "401":
          description: Unauthorized.
          schema:
            type: string
        "403":
          description: Thrown when anyone except the receiver tries to decline a challenge.
          schema:
            type: string
        "500":
          description: Internal Server Error.
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Decline a draft challenge from another user.
      tags:
      - Draft
...
swagger: "2.0"